+++ frontend/package.json
@@ -54,6 +54,7 @@
     "@iconify-json/material-symbols": "^1.2.19",
     "@iconify/json": "^2.2.327",
     "@nuxt/eslint": "^1.4.1",
+    "@nuxt/scripts": "0.11.7",
     "@nuxt/test-utils": "^3.17.2",
     "@nuxtjs/tailwindcss": "^6.14.0",
     "@pact-foundation/pact": "^15.0.1",
@@ -66,6 +67,7 @@
     "@types/jsdom": "^21.1.7",
     "@types/lodash": "^4.17.16",
     "@types/node": "^22.14.1",
+    "@unhead/vue": "^2.0.3",
     "@vitejs/plugin-vue": "^5.2.3",
     "@vitest/coverage-v8": "^3.1.1",
     "@vue/test-utils": "^2.4.6",
