+++ frontend/src/app.vue
@@ -1,11 +1,14 @@
 <script setup lang="ts">
+import { useSentry } from "@/composables/useSentry";
 useHead({
   titleTemplate: (pageTitle) =>
     pageTitle
       ? `${pageTitle} | Rechtsinformationen des Bundes`
       : "Rechtsinformationen des Bundes",
 });
+useSentry();
 </script>
+
 <template>
   <NuxtLayout>
     <NuxtPage />
